R Programming Peer Assignment
R Programming Peer Assignment - Create 2 functions.  makeCacheMatrix()
function to create a matrix which has 4 functions to set and get the
created matrix and to set and get the inverse of the created matrix.
cacheSolve() function is to return the cached inverse matrix if
available or  run a solve() function on the retrieved matrix and set
and return the inverse matrix.